Refactored `Cart` component: moved to `Customer/Cart` directory, updated imports across files; adjusted `Payment` fetch URL, cleaned up console logs in `Payment`, `OrderHistory`, and `Header`, and removed commented-out code in `handleUpdateAddress`.
