+++ quangvinh-store/src/components/layout/customer/Header.jsx
@@ -6,7 +6,7 @@ import { Link } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCartShopping, faHeart, faMagnifyingGlass, faUser } from '@fortawesome/free-solid-svg-icons';
 import { AuthContext } from '../../../context/AuthContext.jsx';
-import Cart from '../../../pages/Customer/Payment/Cart.jsx';
+import Cart from '../../../pages/Customer/Cart/Cart.jsx';
 import Search from '../../../pages/Customer/Product/Search.jsx';
 
 function Header() {
